buildActivityList changed to makeActivityNode
The function now returns an activity node, instead of recursively building a list.  This way we can iterate over the text and possibly check for errors.  The old method was a tail-recursion and was unneccesary.  We can turn this linked list into a queue.  For our purposes the queue is much more appropriate.